pyproject.toml
[project]
name = "frame-stamp"
version = "0.1.12"
description = "Adding technical information to images using a template."
authors = [{ name = "paulwinex", email = "paulwinex@gmail.com" }]
requires-python = ">=3.9,<3.13"
readme = "README.md"
dependencies = [
    "pillow==11.3.0",
    "pyyaml>=6.0.2,<7",
    "recommonmark>=0.7.1,<0.8",
]

[project.optional-dependencies]
pyside = ["pyside6==6.8.1"]
svg = ["cairosvg>=2.7.1,<3"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "sphinx>=7.2.6,<8",
    "sphinx-rtd-theme>=2.0.0,<3",
]

[build-system]
requires = ["uv", "wheel", "setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["frame_stamp*"]
exclude = ["tests*", "*__pycache__*", "_tmp"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"frame_stamp" = [
    "fonts/*",
    "bin/*",
    "resources/*",
]
